backend/requirements.txt
# FastAPI Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
httpx>=0.26.0
redis>=5.0.0
supabase>=2.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
psutil>=5.9.0

# ML & Data
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.4.0
joblib>=1.3.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
email-validator>=2.1.0

# Optional: For development
psycopg2-binary>=2.9.0
